Properly attribute authorship
Summary:
This is a blast from a past that feels so very, very distant now. But still, the ethics of signing what's essentially 60% someone else's work and 100% someone else's idea is... let's say... dubious? There will be a ton of other things to rightfully and exclusively slap your name on, Angelo. But this is not one of them.
Closes https://github.com/facebook/fbtftp/pull/14

Differential Revision: D3360487

Pulled By: pallotron

fbshipit-source-id: 22952cad9de25575c680f25f920a4c94759152d5