Ack TFTP options in request order.
Summary:
The TFTP RFC mentions that "The order in which options are specified is not significant", so this should not be strictly necessary.

HOWEVER: With Python 3.6 dicts are ordered by default, so we inadvertently added an ordering.

At least this way, the ordering is not arbitrary depending on our source but rather the same order as the client requested.
Closes https://github.com/facebook/fbtftp/pull/22

Differential Revision: D4692170

Pulled By: pallotron

fbshipit-source-id: a77459742bedf425561cfe65341f1b9c230dacf2