Fix to handle buggy clients that burst retries
Summary:
Trying to fix the TFTP issue seen at T24419310.

Basically some hosts have buggy TFTP client firmware that send a burst of two TFTP requests with two different source ports. Fbtftp has a bug that is triggered by this condition, i.e. it treats these two requests as duplicate, rathe than as a retry.  So it tries to respond to the original request (with the first source port that was used), but the client does not respond anymore on that port. The fix is to treat the retry as such, rather than as a duplicate.

Differential Revision: D6655906

fbshipit-source-id: e72c6d0657f82d49b146490ff066f1e3338e3370