Fix: README.md - update example (#44)
Summary:
Dear fbtftp maintainers,

First many thanks for this code. I am going to use it in our own project (https://github.com/bluebanquise/bluebanquise) as we rely on the old atftp server that seems no more maintained. Also having a python written tool is a must for us.

During tests of fbtftp, I found that while the example file provided in examples/server.py is working (with a very nice verbosity output !), the example provided in the README.md is not.

Here are the changes done:

* Replace ip by address at StaticServer call
* Replace empty address by '::'
* Replace part as string to integer, and set to 69 default tftp port
* Add missing comma in example

With my best regards

Ox

Pull Request resolved: https://github.com/facebook/fbtftp/pull/44

Reviewed By: lestrozi

Differential Revision: D23959925

Pulled By: pallotron

fbshipit-source-id: fb31b945ed5403f1b5acc1240a1eb4571c88a9a5