Enable Linux and BSD compatibility (#49)
Summary:
Hey guys, first of all, great work on the framework.

I need to use the framework on FreeBSD servers and it was quite a boomer when I saw it was depending on epoll, so I took some time and came up with this solution using `selectors` instead of `select`, that way things got OS agnostic and selectors should in theory use the most efficient method mechanism available.

I took the care of also adjusting the `poll_mock()` in the tests. Please double check the work there as I was slightly unsure if that was the right way to go.

Luiz

Pull Request resolved: https://github.com/facebook/fbtftp/pull/49

Test Plan:
Test logs for Linux and FreeBSD:
https://github.com/facebook/fbtftp/pull/49#issuecomment-860666659

Reviewed By: pallotron

Differential Revision: D29100368

Pulled By: skozlov404

fbshipit-source-id: d5b25392a8891deda6903e65989eef0525be759b